feat: improve messages components organization
