+++ src/components/transaction/messages/gov/MessageVoteListItem.tsx
@@ -3,20 +3,23 @@ import { VoteOption } from 'cosmjs-types/cosmos/gov/v1beta1/gov';
 import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
-import { Typography } from '../../typography';
-import { BaseMessageListItem } from './BaseMessageListItem';
+import { Typography } from '../../../typography';
+import { BaseMessageListItem } from '../BaseMessageListItem';
 
 export type Props = {
-  encodeObject: MsgVoteEncodeObject;
+  message: MsgVoteEncodeObject["value"];
   date: Date;
 };
 
-export const MessageVoteListItem: React.FC<Props> = (props) => {
-  const { encodeObject, date } = props;
+/**
+ * Displays the short details of a MsgVote within a list.
+ * @constructor
+ */
+export const MessageVoteListItem: React.FC<Props> = ({ message, date }) => {
   const { t } = useTranslation();
 
   const voteString = useMemo(() => {
-    switch (encodeObject.value.option) {
+    switch (message.option) {
       case VoteOption.VOTE_OPTION_YES:
         return t('yes');
 
@@ -32,17 +35,17 @@ export const MessageVoteListItem: React.FC<Props> = (props) => {
       default:
         return 'undefined';
     }
-  }, [encodeObject.value.option, t]);
+  }, [message.option, t]);
 
   return (
     <BaseMessageListItem
-      icon={require('../../../assets/tx-icons/vote.png')}
+      icon={require('../../../../assets/tx-icons/vote.png')}
       date={date}
       renderContent={() => (
         <View>
           <Typography.Body1>
             {t('vote proposal number with vote', {
-              number: encodeObject.value.proposalId?.toString() ?? '?',
+              number: message.proposalId?.toString() ?? '?',
               vote: voteString,
             })}
           </Typography.Body1>
