+++ src/components/transaction/messages/profiles/MessageSaveProfileListItem.tsx
@@ -2,21 +2,23 @@ import { MsgSaveProfileEncodeObject } from '@desmoslabs/desmjs';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
-import { Typography } from '../../typography';
-import { BaseMessageListItem } from './BaseMessageListItem';
+import { Typography } from '../../../typography';
+import { BaseMessageListItem } from '../BaseMessageListItem';
 
 export type Props = {
-  encodeObject: MsgSaveProfileEncodeObject;
+  message: MsgSaveProfileEncodeObject["value"];
   date: Date;
 };
 
-export const MessageSaveProfileListItem: React.FC<Props> = (props) => {
-  const { date } = props;
+/**
+ * Displays the short details of a MsgSaveProfile within a list.
+ * @constructor
+ */
+export const MessageSaveProfileListItem: React.FC<Props> = ({message, date}) => {
   const { t } = useTranslation();
-
   return (
     <BaseMessageListItem
-      icon={require('../../../assets/tx-icons/general.png')}
+      icon={require('../../../../assets/tx-icons/general.png')}
       date={date}
       renderContent={() => (
         <View>
