+++ src/components/transaction/messages/profiles/MessageUnlinkChainAccountListItem.tsx
@@ -2,28 +2,30 @@ import { MsgUnlinkChainAccountEncodeObject } from '@desmoslabs/desmjs';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
-import { Typography } from '../../typography';
-import { BaseMessageListItem } from './BaseMessageListItem';
+import { Typography } from '../../../typography';
+import { BaseMessageListItem } from '../BaseMessageListItem';
 
 export type Props = {
-  encodeObject: MsgUnlinkChainAccountEncodeObject;
+  message: MsgUnlinkChainAccountEncodeObject["value"];
   date: Date;
 };
 
-export const MessageUnlinkChainAccountListItem: React.FC<Props> = (props) => {
-  const { date, encodeObject } = props;
-  const { target } = encodeObject.value;
+/**
+ * Displays the short details of a MsgUnlinkChainAccount within a list.
+ * @constructor
+ */
+export const MessageUnlinkChainAccountListItem: React.FC<Props> = ({ message, date }) => {
   const { t } = useTranslation();
 
   return (
     <BaseMessageListItem
-      icon={require('../../../assets/tx-icons/general.png')}
+      icon={require('../../../../assets/tx-icons/general.png')}
       date={date}
       renderContent={() => (
         <View>
           <Typography.Body1>{t('tx type unlink chain account')}</Typography.Body1>
           <View style={{ flexDirection: 'row', flexShrink: 1 }}>
-            <Typography.Caption>{target}</Typography.Caption>
+            <Typography.Caption>{message.target}</Typography.Caption>
           </View>
         </View>
       )}
