+++ src/components/transaction/messages/staking/MessageDelegateListItem.tsx
@@ -3,34 +3,37 @@ import { convertCoin } from '@desmoslabs/desmjs';
 import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
-import useCurrentChainInfo from '../../../hooks/desmosclient/useCurrentChainInfo';
-import { Typography } from '../../typography';
-import { BaseMessageListItem } from './BaseMessageListItem';
+import useCurrentChainInfo from '../../../../hooks/desmosclient/useCurrentChainInfo';
+import { Typography } from '../../../typography';
+import { BaseMessageListItem } from '../BaseMessageListItem';
 
 export type Props = {
-  encodeObject: MsgDelegateEncodeObject;
+  message: MsgDelegateEncodeObject["value"];
   date: Date;
 };
 
-export const MessageDelegateListItem: React.FC<Props> = (props) => {
-  const { encodeObject, date } = props;
+/**
+ * Displays the short details of a MsgDelegate within a list.
+ * @constructor
+ */
+export const MessageDelegateListItem: React.FC<Props> = ({ message, date }) => {
   const { t } = useTranslation();
   const chainInfo = useCurrentChainInfo();
 
   const delegateAmount = useMemo(() => {
-    if (encodeObject.value.amount) {
-      const converted = convertCoin(encodeObject.value.amount, 6, chainInfo.denomUnits);
+    if (message.amount) {
+      const converted = convertCoin(message.amount, 6, chainInfo.denomUnits);
       if (converted !== null) {
         return `${converted.amount} ${converted.denom.toUpperCase()}`;
       }
     }
 
     return '0';
-  }, [encodeObject.value.amount, chainInfo]);
+  }, [message.amount, chainInfo]);
 
   return (
     <BaseMessageListItem
-      icon={require('../../../assets/tx-icons/delegate.png')}
+      icon={require('../../../../assets/tx-icons/delegate.png')}
       date={date}
       renderContent={() => (
         <View>
@@ -39,7 +42,7 @@ export const MessageDelegateListItem: React.FC<Props> = (props) => {
           </Typography.Body1>
           <View style={{ flexDirection: 'row', flexShrink: 1 }}>
             <Typography.Caption>
-              {t('to')} {encodeObject.value.validatorAddress}
+              {t('to')} {message.validatorAddress}
             </Typography.Caption>
           </View>
         </View>
