+++ src/components/transaction/messages/staking/MessageWithdrawDelegatorRewardListItem.tsx
@@ -2,28 +2,27 @@ import { MsgWithdrawDelegatorRewardEncodeObject } from '@cosmjs/stargate';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
-import { Typography } from '../../typography';
-import { BaseMessageListItem } from './BaseMessageListItem';
+import { Typography } from '../../../typography';
+import { BaseMessageListItem } from '../BaseMessageListItem';
 
 export type Props = {
-  encodeObject: MsgWithdrawDelegatorRewardEncodeObject;
+  message: MsgWithdrawDelegatorRewardEncodeObject["value"];
   date: Date;
 };
 
-export const MessageWithdrawDelegatorRewardListItem: React.FC<Props> = (props) => {
-  const { encodeObject, date } = props;
+export const MessageWithdrawDelegatorRewardListItem: React.FC<Props> = ({ message, date }) => {
   const { t } = useTranslation();
 
   return (
     <BaseMessageListItem
-      icon={require('../../../assets/tx-icons/withdraw.png')}
+      icon={require('../../../../assets/tx-icons/withdraw.png')}
       date={date}
       renderContent={() => (
         <View>
           <Typography.Body1>{t('withdraw rewards')}</Typography.Body1>
           <View style={{ flexDirection: 'row', flexShrink: 1 }}>
             <Typography.Caption>
-              {t('from')} {encodeObject.value.validatorAddress}
+              {t('from')} {message.validatorAddress}
             </Typography.Caption>
           </View>
         </View>
