+++ src/components/transaction/messages/unknown/MessageUnknownDetails.tsx
@@ -1,15 +1,19 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
-import { Divider } from '../Divider';
-import { LabeledValue } from '../LabeledValue';
+import { Divider } from '../../../Divider';
+import { LabeledValue } from '../../../LabeledValue';
 
 export type UnknownTxMessageProps = {
   typeUrl: string;
   value: string;
 };
 
-export const UnknownTxMessage: React.FC<UnknownTxMessageProps> = (props) => {
+/**
+ * Displays the full details of an unknown message.
+ * @constructor
+ */
+export const MessageUnknownDetails: React.FC<UnknownTxMessageProps> = (props) => {
   const { typeUrl, value } = props;
   const { t } = useTranslation();
